GptApiTranslate v1.4.0 - Numerous performance improvements and code cleanup
- Switched from curl command line to libcurl library for http requests, thus improving network request efficiency by taking advantage of reusable connections.
- CustomCurlPath config values has been removed, since curl command line is no longer used.
- Leverage Locker module for most mutex functionality
- Switched to more efficient string-wstring conversion process
- Leverage memory cache for config values while no ini file changes detected
- Moved execution requirements functionality to its own module (ExtExecRequirements)
- Store most class code in header files instead of separate cpp files (to minimize the number of code files being maintained)

+++ Textractor.GptApiTranslate/src/Textractor.GptApiTranslate/ExtensionImpl.cpp
@@ -18,7 +18,8 @@ extern "C" __declspec(dllexport) wchar_t* OnNewSentence(wchar_t* sentence, const
 	try
 	{
 		std::wstring sentenceCopy(sentence);
-		int oldSize = sentenceCopy.size();
+		size_t oldSize = sentenceCopy.size();
+
 		if (ProcessSentence(sentenceCopy, SentenceInfo{ sentenceInfo }))
 		{
 			if (sentenceCopy.size() > oldSize) sentence = (wchar_t*)HeapReAlloc(GetProcessHeap(), HEAP_GENERATE_EXCEPTIONS, sentence, (sentenceCopy.size() + 1) * sizeof(wchar_t));
